+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+## 2024-04-04 0.2.0
+
 - Breaking changes to API, `prompt` and `conversation` are now sub commands
 - Deprecated null values in config, please remove.
 - Show better errors when no response is received